+++ src/esm/VEllipse.js
@@ -7,12 +7,15 @@
  * @modified 2021-02-14 Added functions `radiusH` and `radiusV`.
  * @modified 2021-02-26 Added helper function `decribeSVGArc(...)`.
  * @modified 2021-03-01 Added attribute `rotation` to allow rotation of ellipses.
- * @modified 2021-03-04 Added the `vertAt` and `perimeter` functions.
+ * @modified 2021-03-03 Added the `vertAt` and `perimeter` methods.
+ * @modified 2021-03-05 Added the `getFoci`, `normalAt` and `tangentAt` method.
  * @version  1.2.2
  *
  * @file VEllipse
  * @fileoverview Ellipses with a center and an x- and a y-axis (stored as a vertex).
  **/
+import { Line } from "./Line";
+import { Vector } from "./Vector";
 import { Vertex } from "./Vertex";
 import { UIDGenerator } from "./UIDGenerator";
 /**
@@ -116,6 +119,58 @@ export class VEllipse {
     }
     ;
     /**
+     * Get the normal vector at the given angle.
+     * The normal vector is the vector that intersects the ellipse in a 90 degree angle
+     * at the given point (speicified by the given angle).
+     *
+     * Length of desired normal vector can be specified, default is 1.0.
+     *
+     * @method normalAt
+     * @instance
+     * @memberof VEllipse
+     * @param {number} angle - The angle to get the normal vector at.
+     * @param {number=1.0} length - [optional, default=1] The length of the returned vector.
+     */
+    normalAt(angle, length) {
+        const point = this.vertAt(angle);
+        const foci = this.getFoci();
+        // Calculate the angle between [point,focusA] and [point,focusB]
+        const angleA = new Line(point, foci[0]).angle();
+        const angleB = new Line(point, foci[1]).angle();
+        const centerAngle = angleA + (angleB - angleA) / 2.0;
+        const endPointA = point.clone().addX(50).clone().rotate(centerAngle, point);
+        const endPointB = point.clone().addX(50).clone().rotate(Math.PI + centerAngle, point);
+        if (this.center.distance(endPointA) < this.center.distance(endPointB)) {
+            return new Vector(point, endPointB);
+        }
+        else {
+            return new Vector(point, endPointA);
+        }
+    }
+    ;
+    /**
+     * Get the tangent vector at the given angle.
+     * The tangent vector is the vector that touches the ellipse exactly at the given given
+     * point (speicified by the given angle).
+     *
+     * Note that the tangent is just 90 degree rotated normal vector.
+     *
+     * Length of desired tangent vector can be specified, default is 1.0.
+     *
+     * @method tangentAt
+     * @instance
+     * @memberof VEllipse
+     * @param {number} angle - The angle to get the tangent vector at.
+     * @param {number=1.0} length - [optional, default=1] The length of the returned vector.
+     */
+    tangentAt(angle, length) {
+        const normal = this.normalAt(angle, length);
+        // Rotate the normal by 90 degrees, then it is the tangent.
+        normal.b.rotate(Math.PI / 2, normal.a);
+        return normal;
+    }
+    ;
+    /**
      * Get the perimeter of this ellipse.
      *
      * @method perimeter
@@ -131,6 +186,36 @@ export class VEllipse {
         const a = this.radiusH();
         const b = this.radiusV();
         return Math.PI * (3 * (a + b) - Math.sqrt((3 * a + b) * (a + 3 * b)));
+    }
+    ;
+    /**
+     * Get the two foci of this ellipse.
+     *
+     * @method getFoci
+     * @instance
+     * @memberof VEllipse
+     * @return {Array<Vertex>} An array with two elements, the two focal points of the ellipse (foci).
+     */
+    getFoci() {
+        // https://www.mathopenref.com/ellipsefoci.html
+        const rh = this.radiusH();
+        const rv = this.radiusV();
+        const sdiff = rh * rh - rv * rv;
+        // f is the distance of each focs to the center.
+        const f = Math.sqrt(Math.abs(sdiff));
+        // Foci on x- or y-axis?
+        if (sdiff < 0) {
+            return [
+                this.center.clone().addY(f).rotate(this.rotation, this.center),
+                this.center.clone().addY(-f).rotate(this.rotation, this.center)
+            ];
+        }
+        else {
+            return [
+                this.center.clone().addX(f).rotate(this.rotation, this.center),
+                this.center.clone().addX(-f).rotate(this.rotation, this.center)
+            ];
+        }
     }
     ;
     /**